refactor: make pagination in CuiToolbar reusable
