+++ libs/frontend/presentation/codelab-ui/src/layout/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useToolbarPagination.hook'
