+++ libs/frontend/presentation/codelab-ui/src/layout/index.ts
@@ -1,3 +1,4 @@
 export * from './CuiHeader'
 export * from './CuiNavigationBar'
 export * from './CuiSidebar'
+export * from './hooks'
